Fixed UI tests which broke due to changes made to fix (32) and (40)
