+++ ownCloudTests/File List/CreateFolderTests.swift
@@ -13,30 +13,9 @@ import ownCloudMocking
 
 @testable import ownCloud
 
-class CreateFolderTests: XCTestCase {
+class CreateFolderTests: FileTests {
 
 	let hostSimulator: OCHostSimulator = OCHostSimulator()
-
-	override func setUp() {
-		super.setUp()
-	}
-
-	override func tearDown() {
-		super.tearDown()
-		OCMockManager.shared.removeAllMockingBlocks()
-	}
-
-	public typealias OCMRequestCoreForBookmarkCompletionHandler = @convention(block)
-		(_ core: OCCore, _ error: NSError?) -> Void
-
-	public typealias OCMRequestCoreForBookmarkSetupHandler = @convention(block)
-		(_ core: OCCore, _ error: NSError?) -> Void
-
-	public typealias OCMRequestCoreForBookmark = @convention(block)
-		(_ bookmark: OCBookmark, _ setup: OCMRequestCoreForBookmarkSetupHandler, _ completionHandler: OCMRequestCoreForBookmarkCompletionHandler) -> Void
-
-	public typealias OCMRequestChangeSetWithFlags = @convention(block)
-		(_ flags: OCQueryChangeSetRequestFlag, _ completionHandler: OCQueryChangeSetRequestCompletionHandler) -> Void
 
 	/*
 	* PASSED if: Create Folder view is shown
@@ -261,70 +240,5 @@ class CreateFolderTests: XCTestCase {
 		} else {
 			assertionFailure("File list not loaded because Bookmark is nil")
 		}
-	}
-
-	func showFileList(bookmark: OCBookmark, issue: OCIssue? = nil) {
-		if let appDelegate: AppDelegate = UIApplication.shared.delegate as? AppDelegate {
-
-			let query = MockOCQuery(path: "/")
-			let core = MockOCCore(query: query, bookmark: bookmark, issue: issue)
-
-			let rootViewController: MockClientRootViewController = MockClientRootViewController(core: core, query: query, bookmark: bookmark)
-
-			appDelegate.serverListTableViewController?.navigationController?.navigationBar.prefersLargeTitles = false
-			appDelegate.serverListTableViewController?.navigationController?.navigationItem.largeTitleDisplayMode = .never
-			appDelegate.serverListTableViewController?.navigationController?.pushViewController(viewController: rootViewController, animated: true, completion: {
-				appDelegate.serverListTableViewController?.navigationController?.setNavigationBarHidden(true, animated: false)
-			})
-		}
-	}
-
-	func dismissFileList() {
-		if let appDelegate: AppDelegate = UIApplication.shared.delegate as? AppDelegate {
-			appDelegate.serverListTableViewController?.navigationController?.popViewController(animated: false)
-		}
-	}
-
-	// MARK: - Mocks
-	func mockOCoreForBookmark(mockBookmark: OCBookmark) {
-		let completionHandlerBlock : OCMRequestCoreForBookmark = { (bookmark, setupHandler, mockedBlock) in
-			let core = OCCore(bookmark: mockBookmark)
-			setupHandler(core, nil)
-			mockedBlock(core, nil)
-		}
-
-		OCMockManager.shared.addMocking(blocks: [OCMockLocation.ocCoreManagerRequestCoreForBookmark: completionHandlerBlock])
-	}
-
-	func mockQueryPropfindResults(resourceName: String, basePath: String, state: OCQueryState) {
-
-		let completionHandlerBlock : OCMRequestChangeSetWithFlags = { (flags, mockedBlock) in
-
-			var items: [OCItem]?
-
-			let bundle = Bundle.main
-			if let path: String = bundle.path(forResource: resourceName, ofType: "xml") {
-
-				if let data = NSData(contentsOf: URL(fileURLWithPath: path)) {
-					if let parser = OCXMLParser(data: data as Data) {
-						parser.options = ["basePath": basePath]
-						parser.addObjectCreationClasses([OCItem.self])
-						if parser.parse() {
-							items = parser.parsedObjects as? [OCItem]
-						}
-					}
-				}
-			}
-
-			items?.removeFirst()
-
-			let querySet: OCQueryChangeSet = OCQueryChangeSet(queryResult: items, relativeTo: nil)
-			let query: OCQuery = OCQuery()
-			query.state = state
-
-			mockedBlock(query, querySet)
-		}
-
-		OCMockManager.shared.addMocking(blocks: [OCMockLocation.ocQueryRequestChangeSetWithFlags: completionHandlerBlock])
 	}
 }
