+++ ownCloudTests/File List/FileListTests.swift
@@ -13,30 +13,13 @@ import ownCloudMocking
 
 @testable import ownCloud
 
-class FileListTests: XCTestCase {
+class FileListTests: FileTests {
 
 	override func setUp() {
 		super.setUp()
 		OCBookmarkManager.deleteAllBookmarks(waitForServerlistRefresh: true)
 		OCMockManager.shared.removeAllMockingBlocks()
 	}
-
-	override func tearDown() {
-		super.tearDown()
-		OCMockManager.shared.removeAllMockingBlocks()
-	}
-
-	public typealias OCMRequestCoreForBookmarkCompletionHandler = @convention(block)
-		(_ core: OCCore, _ error: NSError?) -> Void
-
-	public typealias OCMRequestCoreForBookmarkSetupHandler = @convention(block)
-		(_ core: OCCore, _ error: NSError?) -> Void
-
-	public typealias OCMRequestCoreForBookmark = @convention(block)
-		(_ bookmark: OCBookmark, _ setup: OCMRequestCoreForBookmarkSetupHandler, _ completionHandler: OCMRequestCoreForBookmarkCompletionHandler) -> Void
-
-	public typealias OCMRequestChangeSetWithFlags = @convention(block)
-		(_ flags: OCQueryChangeSetRequestFlag, _ completionHandler: OCQueryChangeSetRequestCompletionHandler) -> Void
 
 	/*
 	* PASSED if: Disconnect button appears in the view
@@ -95,57 +78,4 @@ class FileListTests: XCTestCase {
 		}
 	}
 
-	func showFileList(bookmark: OCBookmark) {
-		if let appDelegate: AppDelegate = UIApplication.shared.delegate as? AppDelegate {
-			let clientRootViewController = ClientRootViewController(bookmark: bookmark)
-
-			appDelegate.serverListTableViewController?.navigationController?.navigationBar.prefersLargeTitles = false
-			appDelegate.serverListTableViewController?.navigationController?.navigationItem.largeTitleDisplayMode = .never
-			appDelegate.serverListTableViewController?.navigationController?.pushViewController(viewController: clientRootViewController, animated: true, completion: {
-				appDelegate.serverListTableViewController?.navigationController?.setNavigationBarHidden(true, animated: false)
-			})
-		}
-	}
-
-	// MARK: - Mocks
-	func mockOCoreForBookmark(mockBookmark: OCBookmark) {
-		let completionHandlerBlock : OCMRequestCoreForBookmark = { (bookmark, setupHandler, mockedBlock) in
-			let core = OCCore(bookmark: mockBookmark)
-			setupHandler(core, nil)
-			mockedBlock(core, nil)
-		}
-
-		OCMockManager.shared.addMocking(blocks: [OCMockLocation.ocCoreManagerRequestCoreForBookmark: completionHandlerBlock])
-	}
-
-	func mockQueryPropfindResults(resourceName: String, basePath: String, state: OCQueryState) {
-		let completionHandlerBlock : OCMRequestChangeSetWithFlags = { (flags, mockedBlock) in
-
-			var items: [OCItem]?
-
-			let bundle = Bundle.main
-			if let path: String = bundle.path(forResource: resourceName, ofType: "xml") {
-
-				if let data = NSData(contentsOf: URL(fileURLWithPath: path)) {
-					if let parser = OCXMLParser(data: data as Data) {
-						parser.options = ["basePath": basePath]
-						parser.addObjectCreationClasses([OCItem.self])
-						if parser.parse() {
-							items = parser.parsedObjects as? [OCItem]
-						}
-					}
-				}
-			}
-
-			items?.removeFirst()
-
-			let querySet: OCQueryChangeSet = OCQueryChangeSet(queryResult: items, relativeTo: nil)
-			let query: OCQuery = OCQuery()
-			query.state = state
-
-			mockedBlock(query, querySet)
-		}
-
-		OCMockManager.shared.addMocking(blocks: [OCMockLocation.ocQueryRequestChangeSetWithFlags: completionHandlerBlock])
-	}
 }
